consolidation wip
