--- packages/babel-plugin-jsx-dom-expressions/test/__ssr_async_fixtures__/attributeExpressions/code.js
@@ -1,59 +1,0 @@
-const template = (
-  <div id="main" {...results} classList={{ selected: selected }} style={{ color }}>
-    <h1
-      {...results()}
-      disabled
-      title={welcoming()}
-      style={{ "background-color": color(), "margin-right": "40px" }}
-      classList={{ selected: selected() }}
-    >
-      <a href={"/"} ref={link}>
-        Welcome
-      </a>
-    </h1>
-  </div>
-);
-
-const template2 = (
-  <div>
-    <div textContent={rowId} />
-    <div textContent={row.label} />
-  </div>
-);
-
-const template3 = (
-  <div
-    id={/*@once*/ state.id}
-    style={/*@once*/ { "background-color": state.color }}
-    name={state.name}
-    textContent={/*@once*/ state.content}
-  />
-);
-
-const template4 = <div class="hi" className={state.class} />;
-
-const template5 = <div class="a" className="b"></div>;
-
-const template6 = <div style={someStyle()} textContent="Hi" />;
-
-const template7 = (
-  <div
-    style={{ "background-color": color(), "margin-right": "40px", ...props.style }}
-    style:padding-top={props.top}
-  />
-);
-
-let refTarget;
-const template8 = <div ref={refTarget} />;
-
-const template9 = <div ref={e => console.log(e)} />;
-
-const template10 = <div ref={refFactory()} />;
-
-const template11 = <div use:something use:another={thing} />;
-
-const template12 = <div prop:htmlFor={thing} />;
-
-const template13 = <input type="checkbox" checked={true} />;
-
-const template14 = <input type="checkbox" checked={state.visible} />;
