--- packages/babel-plugin-jsx-dom-expressions/test/__ssr_async_fixtures__/simpleElements/code.js
@@ -1,9 +1,0 @@
-const template = (
-  <div id="main">
-    <style>{"div { color: red; }"}</style>
-    <h1>Welcome</h1>
-    <label for={"entry"}>Edit:</label>
-    <input id="entry" type="text" />
-    {/* Comment Node */}
-  </div>
-);
