--- packages/babel-plugin-jsx-dom-expressions/test/__ssr_async_fixtures__/simpleElements/output.js
@@ -1,5 +1,0 @@
-import { ssr as _$ssr } from "r-server";
-
-const template = _$ssr(
-  '<div id="main"><style>div { color: red; }</style><h1>Welcome</h1><label for="entry">Edit:</label><input id="entry" type="text"></div>'
-);
