--- packages/dom-expressions/src/ssr.js
@@ -1,119 +1,0 @@
-import { Aliases, BooleanAttributes } from "./constants";
-export { effect, memo, currentContext, createComponent } from "rxcore";
-export { assignProps, dynamicProperty, getHydrationKey } from "./shared";
-
-export function ssrClassList(value) {
-  let classKeys = Object.keys(value),
-    result = "";
-  for (let i = 0, len = classKeys.length; i < len; i++) {
-    const key = classKeys[i],
-      classValue = !!value[key];
-    if (!key || !classValue) continue;
-    i && (result += " ");
-    result += key;
-  }
-  return result;
-}
-
-export function ssrStyle(value) {
-  if (typeof value === "string") return value;
-  let result = "";
-  const k = Object.keys(value);
-  for (let i = 0; i < k.length; i++) {
-    const s = k[i];
-    if (i) result += ";";
-    result += `${s}:${escape(value[s], true)}`;
-  }
-  return result;
-}
-
-export function ssrSpread(props) {
-  return () => {
-    if (typeof props === "function") props = props();
-    // TODO: figure out how to handle props.children
-    const keys = Object.keys(props);
-    let result = "";
-    for (let i = 0; i < keys.length; i++) {
-      const prop = keys[i];
-      if (prop === "children") {
-        console.warn(`SSR currently does not support spread children.`);
-        continue;
-      }
-      const value = props[prop];
-      if (prop === "style") {
-        result += `style="${ssrStyle(value)}"`;
-      } else if (prop === "classList") {
-        result += `class="${ssrClassList(value)}"`;
-      } else if (BooleanAttributes.has(prop)) {
-        if (value) result += prop;
-        else continue;
-      } else {
-        result += `${Aliases[prop] || prop}="${escape(value, true)}"`;
-      }
-      if (i !== keys.length - 1) result += " ";
-    }
-    return result;
-  };
-}
-
-export function ssrBoolean(key, value) {
-  return value ? " " + key : "";
-}
-
-export function escape(s, attr) {
-  const t = typeof s;
-  if (t !== "string") {
-    if (attr && t === "boolean") return String(s);
-    return s;
-  }
-  const delim = attr ? '"' : "<";
-  const escDelim = attr ? "&quot;" : "&lt;";
-  let iDelim = s.indexOf(delim);
-  let iAmp = s.indexOf("&");
-
-  if (iDelim < 0 && iAmp < 0) return s;
-
-  let left = 0,
-    out = "";
-
-  while (iDelim >= 0 && iAmp >= 0) {
-    if (iDelim < iAmp) {
-      if (left < iDelim) out += s.substring(left, iDelim);
-      out += escDelim;
-      left = iDelim + 1;
-      iDelim = s.indexOf(delim, left);
-    } else {
-      if (left < iAmp) out += s.substring(left, iAmp);
-      out += "&amp;";
-      left = iAmp + 1;
-      iAmp = s.indexOf("&", left);
-    }
-  }
-
-  if (iDelim >= 0) {
-    do {
-      if (left < iDelim) out += s.substring(left, iDelim);
-      out += escDelim;
-      left = iDelim + 1;
-      iDelim = s.indexOf(delim, left);
-    } while (iDelim >= 0);
-  } else
-    while (iAmp >= 0) {
-      if (left < iAmp) out += s.substring(left, iAmp);
-      out += "&amp;";
-      left = iAmp + 1;
-      iAmp = s.indexOf("&", left);
-    }
-
-  return left < s.length ? out + s.substring(left) : out;
-}
-
-export function resolveSSRNode(node) {
-  const t = typeof node;
-  if (t === "string") return node;
-  if (node == null || t === "boolean") return "";
-  if (Array.isArray(node)) return node.map(resolveSSRNode).join("");
-  if (t === "object") return resolveSSRNode(node.t);
-  if (t === "function") return resolveSSRNode(node());
-  return String(node);
-}
