+++ packages/dom-expressions/test/insert.spec.js
@@ -1,4 +1,4 @@
-import * as r from '../src/runtime';
+import * as r from '../src/client';
 import * as S from "s-js";
 
 describe("r.insert", () => {
