+++ packages/lit-dom-expressions/test/html.spec.js
@@ -1,6 +1,6 @@
 import * as S from "s-js";
 import { createHTML } from "../dist/lit-dom-expressions";
-import * as r from "dom-expressions/src/runtime";
+import * as r from "dom-expressions/src/client";
 
 const html = createHTML(r);
